Tests/LibWeb/Text/input/css/animation-name-is-valid-other-property-value.html

<!DOCTYPE html>
<html>
    <body>
        <style>
            #timing-function {
                animation: none;
                animation-name: eAsE-In-OuT;
            }
            #direction {
                animation: none;
                animation-name: rEvErSe;
            }
            #fill-mode {
                animation: none;
                animation-name: fOrWaRdS;
            }
            #running-state {
                animation: none;
                animation-name: pAuSeD;
            }
        </style>
        <script src="../include.js"></script>
        <script>
            test(() => {
                for (const rule of document.styleSheets[0].cssRules) {
                    println(rule.cssText);
                }
            });
        </script>
    </body>
</html>
