Tests/LibWeb/Text/expected/css/animation-name-is-valid-other-property-value.txt

#timing-function { animation: ease eAsE-In-OuT; }
#direction { animation: normal rEvErSe; }
#fill-mode { animation: none fOrWaRdS; }
#running-state { animation: running pAuSeD; }
